requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==36.1.1
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.4
pytest-django==4.10.0
pytest-factoryboy==2.7.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
